Change of strategy: Previously, I started to write semantic actions directly in grammar in order to do all the parsing steps (syntax and semantic analysis) at the same time. I felt forced to do that because of possible ambiguities such as: 	a < b || c > d; //is 'a' a boolean value or a class template type? Semantic actions of the grammar would perform the semantic analysis which results were used to do the disambiguations required in the above case. Obviously, the grammar was going to be dirty, and we were going to break the single responsibility principle.
I finally found a way to do the parsing work in different and clearly
separated modules.
The first step after the preprocessing is a preliminary syntax analysis
which only analyses declarations. The statements (i.e. function's bodies)
are simply 'skipped' by it. The AST of this analysis will be processed by
a semantic analyzer which will build a partial model of the program. These
semantic informations will finally be used by a final syntax analyzer
which will finalize the AST. Then a final semantic analyzer will finalize
the model of the program.

The grammar of the preliminary syntax analyzer is wrote. The latter is
able to parse the whole standard library header set.
Hurray.

Code needs to be refactorised.