Finally, resolving ambiguous expressions by postponing function's bodies analysis won't be enough, since such expressions can be found out of a function's body (see test/testfiles/single_files/test030.cpp). Some minimal semantic analysis have to be performed during syntax analysis (for name lookup).
E.g.:
	a < b || c > d;
Here, we have to know whether 'a' is a type name to be able to know what
kind of expression is this (boolean expression or template type
declaration).

Consequently:
	- declaration_*_analyzer classes become *_analyzer
	- spirit's semantic actions (of some previous revision) are
	  reinserted
	- so will be spirit's functor parsers

These ar bad news, since I was pretty sure that my strategy were good.
However, this is not a big deal. Indeed, no big changes had to be done.